WIP try to optimize using priority queue
Does not yet work. Slower and incorrect.
The idea is to visit each new position exactly once coming from the least risky path. This enables a global visibility map. The way by which we visit a new position by the least risky path is a priority queue which always has the lowest risk path as its first item. So when exploring to new positions from this path, it should be the least risky path to that new position.